Implement contact detail retrieval API; move GET logic to new route file
