+++ src/app/api/contact/route.js
@@ -37,26 +37,6 @@ export async function GET() {
   }
 }
 
-// GET - Lấy chi tiết Contact theo ID
-export async function GET(request, { params }) {
-  await dbConnect();
-
-  try {
-    const contact = await Contact.findById(params.id);
-    if (!contact) {
-      return new Response(JSON.stringify({ error: "Contact not found" }), {
-        status: 404,
-      });
-    }
-    return new Response(JSON.stringify(contact), { status: 200 });
-  } catch (error) {
-    console.error("Error fetching contact:", error);
-    return new Response(JSON.stringify({ error: "Failed to fetch contact" }), {
-      status: 500,
-    });
-  }
-}
-
 // DELETE - Xoá Contact theo ID
 export async function DELETE(request, { params }) {
   await dbConnect();
